gridtrader-pro/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
alembic==1.13.0
yfinance==0.2.22
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
websockets==12.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
authlib==1.2.1
httpx==0.25.2
starlette==0.27.0
supervisor==4.2.5
